TASK: Improve export when form definition changes, allow setting nodeTypesIgnoredInExport
This package only stores the FormElement identifier at the time of form
submission in its entry data. This has some drawbacks:

* If a field identifier is added or removed later, it breaks the export
* The field identifier could be a uuid which makes it hard to read

This change improves this as follows:

* A unique list of each field identifier of each submission is compiled
* The ContentRepository is used to look up the actual field labels if
  the form still exists.
* There is a fallback to the field identifier if the form or field don't
  exist anymore.

Furthermore, the export contains too much information, such as Sections,
Captchas (if configured) and possibly sensible information such as
Passwords.

A new configuration nodeTypesIgnoredInExport is added to prevent the output
of fields configured. The following types are excluded by default:

- Neos.Form.Builder:Section
- Neos.Form.Builder:StaticText
- Neos.Form.Builder:Password
- Neos.Form.Builder:PasswordWithConfirmation