TASK: Add setting to ignore values of certain form element types in Finisher
It may not be desirable to store data for certain form element types
in a Node-based form. For example, for security reasons you might not
want to save the content of Password fields or you might want to
skip saving a captcha value.

To achieve this, a new configuration "nodeTypesIgnoredInFinisher"
is added, by default containing

- Neos.Form.Builder:Section
- Neos.Form.Builder:StaticText