fix: Check multiple dimensions to find matching form
This feature will allow sites with one or multiple dimensions to place the form in a dimension that is not the default dimension. Previously no form could be found in such case.
This is done by iterating over all dimensions and use the first dimension that has a valid form for the export. Remaining dimensions are ignored.