BUGFIX: Define query parameter as array
The problem in is that the IN() clause in SQL expects a list of values, but the code currently passes a single string containing all the values ​​(e.g. "value1,value2,value3") instead of an actual one List of values ​​(['value1', 'value2', 'value3']). So we use now the type PARAM_STR_ARRAY.